Move theming tests to commonTest
